There are primarily three things being done here.
1. The Packet/Particle structure had a raw pointer from which we were
   making slices and handing it out to applications that need to read
   or write to/from the packet. There were concerns (as seen in the
   issue details) that its not done the proper Rust way, and someone
   had suggested replacing the pointer in the struct with an &[]  slice
   and introducing proper lifetimes. So thats the primary change here,
   the raw pointer is now a slice, and it has a static lifetime. See
   Docs/lifetimes.md for details on that.

2. There was a comment about not using global packet pools and having
   per thread pools. I was not convinced about it in the beginning, but
   some more thinking about it has swayed me in favour of the idea, so
   now I have per thread pools configured. So the packets that are handed
   off to other threads are freed back into a queue (one per thread) and
   the thread that owns that packet will pick it up from the queue and
   free it back to its own pool. See packet/packet.md for details

3. There was another comment about explicitly passing String where needed
   rather than passing &str and hiding the internal conversion to String,
   taken care of that in few places

Things not taken care of: the reddit comment thread also talked about freezing
the packet into a read-only structure before handing off to another thread. My
code already works in such a fashion that a packet is completely formed (written)
inside one thread and the handoff is purely done to send it out of a driver that
is handled by another thread - so its already in a mode where the other thread
just reads the packet. Enforcing the read-only via structure changes is not
hard, I plan to do it in a future commit

Fixes: #16