Use shared memory counter array to implement perf counters
An array of three counters is used. The first one is used to store the rdtsc()
value. The second is the number of times this perf point was hit, the third is
the cumulative of all the deltas measured by the perf point. The perf point is
basically a perf.start() and perf.stop() - the time between that is the delta.
So over time, the sum-of-deltas / hit-count gives an average ticks-taken by the
code executed in that perf point. Right now the perf counters are all dumped
by the r2cnt utility, the perf counter have a "perf" prefix on which to grep
for, at some point I will make an r2cnt option just to see perf counters.

Also bunching along with this commit a change to make all the counter flavours
take mutable parameters for the get/set/add/incr options, for some reason I
had not made it mutable before, but it really is mutable since we are mutating
them.

fixes #14